requirements.txt
Flask==3.0.0
openstacksdk==1.0.0
waitress==2.1.2
keystoneauth1==5.2.0
